split up spec/support
